+++ frontend/src/components/atoms/index.ts
@@ -2,4 +2,5 @@ export { default as Page } from "./Page";
 export { default as FloatingCard } from "./FloatingCard";
 export { default as GraphKey } from "./GraphKey";
 export { default as GraphResetButton } from "./GraphResetButton";
+export { default as GraphHideEdgesButton } from "./GraphHideEdgesButton";
 export { default as InstanceType } from "./InstanceType";